+++ mobile/lib/screens/profile_screen.dart
@@ -5,6 +5,7 @@ import '../config/app_version.dart';
 import '../providers/auth_provider.dart';
 import 'auth/login_screen.dart';
 import 'notification_screen.dart';
+import 'settings_screen.dart';
 
 /// 내정보 화면
 /// - 비회원: 로그인 안내 표시
@@ -193,15 +194,18 @@ class _ProfileScreenState extends ConsumerState<ProfileScreen> {
                   },
                 ),
                 _buildDivider(),
-                // 설정 메뉴 (향후 사용 예정)
-                // _buildMenuItem(
-                //   icon: Icons.settings_outlined,
-                //   title: '설정',
-                //   onTap: () {
-                //     // TODO: 설정 화면으로 이동
-                //   },
-                // ),
-                // _buildDivider(),
+                _buildMenuItem(
+                  icon: Icons.settings_outlined,
+                  title: '설정',
+                  onTap: () {
+                    Navigator.of(context).push(
+                      MaterialPageRoute(
+                        builder: (context) => const SettingsScreen(),
+                      ),
+                    );
+                  },
+                ),
+                _buildDivider(),
                 _buildMenuItem(
                   icon: Icons.logout,
                   title: '로그아웃',
@@ -227,25 +231,6 @@ class _ProfileScreenState extends ConsumerState<ProfileScreen> {
                 fontSize: 12.sp,
                 fontWeight: FontWeight.w400,
                 color: const Color(0xFF9CA3AF),
-              ),
-            ),
-          ),
-
-          SizedBox(height: 16.h),
-
-          // 회원 탈퇴 버튼
-          Center(
-            child: TextButton(
-              onPressed: () => _showDeleteAccountBottomSheet(context),
-              child: Text(
-                '회원 탈퇴',
-                style: TextStyle(
-                  fontSize: 12.sp,
-                  fontWeight: FontWeight.w400,
-                  color: const Color(0xFF9CA3AF),
-                  decoration: TextDecoration.underline,
-                  decorationColor: const Color(0xFF9CA3AF),
-                ),
               ),
             ),
           ),
@@ -432,236 +417,6 @@ class _ProfileScreenState extends ConsumerState<ProfileScreen> {
             ),
           ),
         ],
-      ),
-    );
-  }
-
-  /// 회원 탈퇴 바텀시트 (요즘 스타일)
-  void _showDeleteAccountBottomSheet(BuildContext context) {
-    final reasonController = TextEditingController();
-    bool isLoading = false;
-
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: Colors.transparent,
-      builder: (context) => StatefulBuilder(
-        builder: (context, setState) => Container(
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.vertical(
-              top: Radius.circular(20.r),
-            ),
-          ),
-          padding: EdgeInsets.only(
-            bottom: MediaQuery.of(context).viewInsets.bottom,
-          ),
-          child: SafeArea(
-            child: Padding(
-              padding: EdgeInsets.all(24.w),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  // 핸들 바
-                  Center(
-                    child: Container(
-                      width: 40.w,
-                      height: 4.h,
-                      decoration: BoxDecoration(
-                        color: const Color(0xFFE5E7EB),
-                        borderRadius: BorderRadius.circular(2.r),
-                      ),
-                    ),
-                  ),
-                  SizedBox(height: 24.h),
-
-                  // 타이틀
-                  Text(
-                    '정말 탈퇴하시겠어요?',
-                    style: TextStyle(
-                      fontSize: 20.sp,
-                      fontWeight: FontWeight.w700,
-                      color: const Color(0xFF1A1C1E),
-                    ),
-                  ),
-                  SizedBox(height: 12.h),
-
-                  // 안내 문구
-                  Text(
-                    '탈퇴하시면 모든 데이터가 삭제되며,\n복구할 수 없습니다.',
-                    style: TextStyle(
-                      fontSize: 14.sp,
-                      fontWeight: FontWeight.w400,
-                      color: const Color(0xFF6C7278),
-                      height: 1.5,
-                    ),
-                  ),
-                  SizedBox(height: 24.h),
-
-                  // 탈퇴 사유 입력 (선택)
-                  Text(
-                    '탈퇴 사유 (선택)',
-                    style: TextStyle(
-                      fontSize: 14.sp,
-                      fontWeight: FontWeight.w500,
-                      color: const Color(0xFF1A1C1E),
-                    ),
-                  ),
-                  SizedBox(height: 8.h),
-                  TextField(
-                    controller: reasonController,
-                    maxLines: 3,
-                    maxLength: 200,
-                    decoration: InputDecoration(
-                      hintText: '서비스 개선을 위해 탈퇴 사유를 알려주세요',
-                      hintStyle: TextStyle(
-                        fontSize: 14.sp,
-                        color: const Color(0xFF9CA3AF),
-                      ),
-                      filled: true,
-                      fillColor: const Color(0xFFF9FAFB),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12.r),
-                        borderSide: const BorderSide(
-                          color: Color(0xFFE5E7EB),
-                        ),
-                      ),
-                      enabledBorder: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12.r),
-                        borderSide: const BorderSide(
-                          color: Color(0xFFE5E7EB),
-                        ),
-                      ),
-                      focusedBorder: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(12.r),
-                        borderSide: BorderSide(
-                          color: Theme.of(context).primaryColor,
-                        ),
-                      ),
-                      contentPadding: EdgeInsets.all(16.w),
-                    ),
-                  ),
-                  SizedBox(height: 24.h),
-
-                  // 버튼들
-                  Row(
-                    children: [
-                      // 취소 버튼
-                      Expanded(
-                        child: OutlinedButton(
-                          onPressed: isLoading
-                              ? null
-                              : () => Navigator.of(context).pop(),
-                          style: OutlinedButton.styleFrom(
-                            foregroundColor: const Color(0xFF6C7278),
-                            side: const BorderSide(
-                              color: Color(0xFFE5E7EB),
-                            ),
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(12.r),
-                            ),
-                            padding: EdgeInsets.symmetric(vertical: 16.h),
-                          ),
-                          child: Text(
-                            '취소',
-                            style: TextStyle(
-                              fontSize: 16.sp,
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                        ),
-                      ),
-                      SizedBox(width: 12.w),
-                      // 탈퇴 버튼
-                      Expanded(
-                        child: ElevatedButton(
-                          onPressed: isLoading
-                              ? null
-                              : () async {
-                                  setState(() => isLoading = true);
-
-                                  try {
-                                    await ref
-                                        .read(authProvider.notifier)
-                                        .deleteAccount(
-                                          reason: reasonController.text.isEmpty
-                                              ? null
-                                              : reasonController.text,
-                                        );
-
-                                    if (context.mounted) {
-                                      Navigator.of(context).pop();
-                                      ScaffoldMessenger.of(context).showSnackBar(
-                                        SnackBar(
-                                          content: const Text(
-                                            '회원 탈퇴가 완료되었습니다.\n그동안 이용해 주셔서 감사합니다.',
-                                          ),
-                                          backgroundColor: Colors.green,
-                                          behavior: SnackBarBehavior.floating,
-                                          shape: RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(10.r),
-                                          ),
-                                        ),
-                                      );
-                                    }
-                                  } catch (e) {
-                                    setState(() => isLoading = false);
-                                    if (context.mounted) {
-                                      ScaffoldMessenger.of(context).showSnackBar(
-                                        SnackBar(
-                                          content: Text(
-                                            e.toString().replaceAll(
-                                                'Exception: ', ''),
-                                          ),
-                                          backgroundColor: Colors.red,
-                                          behavior: SnackBarBehavior.floating,
-                                          shape: RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(10.r),
-                                          ),
-                                        ),
-                                      );
-                                    }
-                                  }
-                                },
-                          style: ElevatedButton.styleFrom(
-                            backgroundColor: Colors.red,
-                            foregroundColor: Colors.white,
-                            elevation: 0,
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(12.r),
-                            ),
-                            padding: EdgeInsets.symmetric(vertical: 16.h),
-                          ),
-                          child: isLoading
-                              ? SizedBox(
-                                  width: 20.w,
-                                  height: 20.w,
-                                  child: const CircularProgressIndicator(
-                                    strokeWidth: 2,
-                                    valueColor: AlwaysStoppedAnimation<Color>(
-                                      Colors.white,
-                                    ),
-                                  ),
-                                )
-                              : Text(
-                                  '탈퇴하기',
-                                  style: TextStyle(
-                                    fontSize: 16.sp,
-                                    fontWeight: FontWeight.w600,
-                                  ),
-                                ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
       ),
     );
   }
